게시물=유저 관계설정 및 게시물 API 수정

+++ src/main/java/com/example/board/service/PostService.java
@@ -6,7 +6,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.example.board.exception.post.PostNotFoundException;
+import com.example.board.exception.user.UserNotAllowedException;
 import com.example.board.model.entity.PostEntity;
+import com.example.board.model.entity.UserEntity;
 import com.example.board.model.post.Post;
 import com.example.board.model.post.PostPatchRequestBody;
 import com.example.board.model.post.PostPostRequestBody;
@@ -31,27 +33,34 @@ public class PostService {
         return Post.from(postEntity);
     }
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        var postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-        var savedPostEntity = postEntityRepository.save(postEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+        var postEntity = postEntityRepository.save(PostEntity.of(postPostRequestBody.body(), currentUser));
 
-        return Post.from(savedPostEntity);
+        return Post.from(postEntity);
     }
 
-    public Post updatePost(PostPatchRequestBody postPatchRequestBody, Long postId) {
+    public Post updatePost(PostPatchRequestBody postPatchRequestBody, Long postId, UserEntity curreUserEntity) {
         var postEntity = postEntityRepository.findById(postId)
                             .orElseThrow(() -> new PostNotFoundException(postId));
+        
+        if(!postEntity.getUser().equals(curreUserEntity)){
+            throw new UserNotAllowedException();
+        }
+
         postEntity.setBody(postPatchRequestBody.body());
         var updatedPostEntity = postEntityRepository.save(postEntity);
 
         return Post.from(updatedPostEntity);
     }
 
-    public void deletePost(Long postId) {
-
+    public void deletePost(Long postId, UserEntity curreUserEntity) {
         var postEntity = postEntityRepository.findById(postId)
                             .orElseThrow(() -> new PostNotFoundException(postId));
+
+        if(!postEntity.getUser().equals(curreUserEntity)){
+            throw new UserNotAllowedException();
+        }
+    
         postEntityRepository.delete(postEntity);
     }
